+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Do not blow up if `diffgram` in XML response is missing. Return nil instead.
+
 ## [0.1.0] - 2016-12-31
 ### Added
 - Initial release